requirements.txt
pandas
sib-api-v3-sdk
python-dotenv
flask
requests
